requirements.txt
# Core framework
streamlit>=1.28.0

# Google Gemini LLM integration
google-generativeai>=0.3.0

# Data visualization
plotly>=5.17.0
pandas>=2.0.0

# QR Code generation
qrcode[pil]>=7.4.0

# Additional dependencies (installed automatically with above)
# - Pillow (for QR code image generation)
# - numpy (for pandas)
# - sqlite3 (built-in with Python)
